Update patch set 5
Patch Set 5:

> I would've used math/big.ErrNaN if its msg field was exported

This is a good point. As I understand it, we plan to migrate
exp/decimal to math/big later? And if that happens, it will
have access to math/big.ErrNaN. I'm wondering if we have a
way to handle this nicely so that users don't need to change
their code to switch to a new error.

Patch-set: 5