Update patch set 5
Patch Set 5:

> > I would've used math/big.ErrNaN if its msg field was exported
 > 
 > This is a good point. As I understand it, we plan to migrate
 > exp/decimal to math/big later? And if that happens, it will
 > have access to math/big.ErrNaN. I'm wondering if we have a
 > way to handle this nicely so that users don't need to change
 > their code to switch to a new error.

I would love to have exp/decimal migrated to math/big. I think Go would benefit from a decimal type in the stdlib, be it fixed- or arbitrary-precision.

It's not a breaking change to export math/big.ErrNaN's msg field, is it? That would solve the problem. The only other reasonably alternative I can think of is `NewErrNaN(string) ErrNaN' but that seems a bit overkill.

Patch-set: 5