Update patch set 4
Patch Set 4:

> My working set of vector images is the Material Design icons. I'll
 > work on getting a histogram of encoding lengths.

Preliminary data from converting the Material Design icons (961 SVG files, 312887 bytes) to IconVG:

The number of 1, 2 and 4 byte encoded coordinates are 39942, 1311 and 25150, totalling 143164 bytes. The total IconVG set (all up, not just the coordinates) weigh in at 172566 bytes.

If we restrict the encoder to round the coordinates to the 1/64 grid (or 3072 units in the equivalent of an em-square, which is finer resolution than a typical TrueType font's 2048 units):

The number of 1, 2 and 4 byte encoded coordinates are 39942, 26461 and 0, totalling 92864 bytes. The total IconVG set (all up, not just the coordinates) weigh in at 122266 bytes.

Patch-set: 4