shiny/widget: a tabbed container
By default it looks a little ugly with a tiny fixed-width font on
a high-DPI screen, but the fix for that is a prettier font.

This also has a reasonably crude version of scrolling in the tab
header that I hope can be done by a widget.Sheet eventually. But for
all its messiness it works.

I've left all the animation logic out of this CL because it is even
messier and should probably wait until the Sheet change is done.

Change-Id: I0b3da0188560b1cdccebcb6b699534608069dc1b