Update patch set 1
Patch Set 1:

I'm not really sure if this should be reviewed now or not. It works, which is nice. But I suspect it will change a lot when more scrolling work is merged in.

But I had it lying around and thought it at least belonged in a CL.

Patch-set: 1
Reviewer: Gerrit User 5030 <5030@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0