shiny/driver/gldriver: correct EGL client context version for x11
EGL documentation at
https://www.khronos.org/registry/EGL/sdk/docs/man/html/eglCreateContext.xhtml
suggest that the correct value for the OpenGL ES 2 context is '2' instead of '3'.
The latter does not seem to be valid according to the specifications.

When using this driver with the previous context version, one of my older test
machines fails with: 'eglCreateContext failed: EGL_BAD_CONFIG'

Change-Id: Iab49576dd2fe7672f969ae5ce74dbb5ad518374e