shiny/driver: add Plan 9 driver
The new Plan 9 driver uses files in the /dev filesystem in order to
convert events that come over the filesystem into shiny events.

It uses a goroutine to continuously read /dev/cons and convert the runes
that it reads into keyboard events, while another goroutine monitors
/dev/mouse to convert the events into shiny mouse or window movement
events.

This has some limitations, such as the fact that we don't know for
sure how the rune was generated. For instance was the rune '\t'
generated by Ctrl-I or the tab key? How do we tell the difference
between Ctrl-H and backspace? In practice, however, this works well
enough for most use cases.

The driver keeps track of the active window, and draws over it using
/dev/draw (emulating the way that most Plan 9 graphical programs work),
but a side-effect of this is that multi-window shiny programs will
likely not currently work.

Change-Id: Ia777d9ac57549a4298222ddf2c15345843fcfec2