shiny/driver/gldriver: be less racy wrt size events.
Set the windowImpl.sz cached value on the receiver side of the channel
communication, not spawned by a goroutine 'soon' after the send.

Also remove what I suspect is a superfluous paint.Event on the Win32
flavor of the gldriver. The other flavors don't do this.

Also fix the non-superfluous paint.Event's External field to true, the
same as the other flavors (cocoa and x11).

Change-Id: Ibe4bad965fe0c059378daef2f535dcae362447a9