shiny/driver/internal/event: use a channel in Deque instead of a sync.Cond
This eliminates spurious wakeups in NextEvent.

sync.Cond allows some other goroutine calling NextEvent to barge in
and take the event in between the call to Signal and the corresponding
return from Wait. In contrast, the channel transmits the event itself
along with the wakeup signal, so if some other goroutine barges in the
worst it can do is update the head event before the waiter can
reacquire the mutex.

In the context of Shiny itself the spurious wakeups are unlikely to
matter much, but since Shiny is in golang.org/x it sets an example
that others are likely to emulate.

updates golang/go#21165

Change-Id: I3585d89b00f532fbc955da23b9dd0e5e6aebd450