shiny/driver/internal/event: eliminate steady-state memory allocations
The current implementation uses a rolling slice for the back of the
queue and a stack-like slice for the front. That still gives
amortized-O(1) behavior, but puts unnecessary pressure on the garbage
collector.

This change removes the rolling behavior for the back slice, instead
copying it (in reverse order) to the front slice when the front
becomes empty. The resulting behavior is that, in the steady state,
the front and back slices converge to some stable upper bound on the
length of the queue and then perform no further allocations.

Change-Id: I29b0dae0e438bb48853796ff79433fc0f7d83b19