shiny/driver/internal/event: use a channel in Deque instead of a sync.Cond
This change is a simpler alternative to https://golang.org/cl/94138.
It differs in that it does not attempt to avoid spurious wakeups, and
therefore does not need to send the queue elements on the same channel
as the wakeup signals.

Since Shiny is in golang.org/x, it sets an example that others are
likely to emulate. The pattern shown in this change adapts more easily
to fit other Go concurrency patterns: for example, in this version of
NextEvent it would be straightforward to select on cancellation of a
context.Context.

This change also makes the locking invariants more explicit: every
Lock and Unlock operation appears in the code (rather than indirectly
through sync.Cond.Signal), so it is more obvious that NextEvent
reacquires q.mu (and must recheck the condition) after receiving the
signal.

Change-Id: If6ceba6c41706f78b6fc6c5ffb2f28faeedf3129