exp/errors: new package
This package implements a version of the Go 2 draft designs for error
inspection and printing.

The Format method has been renamed ErrorFormatter, to allow errors to
support the new formatting behavior without changes to the fmt package.
An errors.Format function allows error to adopt this formatting
with minimal effort.

This package includes additional features not fully covered in the
draft designs. All errors include a single call frame of stack
information. Errors may implement Is and As methods to support rich
matching and impersonation respectively.

Change-Id: I57736959b9918b384efa93a7b3a13df44d05ca53