Update patch set 2
Patch Set 2:

I support introducing a golang.org/x/exp/errors, but it needs to be a drop-in replacement for the standard 'errors' package, with the eventual goal of moving the new stuff to standard 'errors' in a way that the standard 'errors' becomes a drop-in replacement for this package.

Let's split this into multiple CLs. The first should just create a workalike for the current errors (delegating to current errors). Then let's add Unwrap/Is/As in a second CL. Then formatting (probably done by Marcel) in a third.

I am OK with having a go.mod here since this is exp, but it should not depend on anything else. I haven't looked through the code enough to see why it mentions x/tools but let's please drop that, whatever it is.

Patch-set: 2
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>