Update patch set 2
Patch Set 2:

> Patch Set 2:

> Any thoughts on putting this in a new repo (golang.org/x/errors, maybe)? I find Dave's arguments against exp compelling.
I strongly prefer x/exp.

I understand Dave's argument, but ideally this should be a relatively short-lived package; most of this stuff can make it in core in the not-too distant future.

Putting it in a separate repo makes it too official and has the danger that it starts leading its own live too much, making it harder to get rid off (cf. context).

It would be nice, though, if people could add "replace" entries in their modules to allow replacing uses of "errors" and "fmt" to these drop-in replacements. But I reckon that only requires us to put a go.mod file in x/errors to make that possible.

Patch-set: 2
Reviewer: Gerrit User 5182 <5182@62eb7196-b449-3ce5-99f1-c037f21e1705>