Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 2:
> 
> (3 comments)
> 
> I'm breaking this up into smaller CLs (starting with https://go-review.googlesource.com/c/exp/+/137795); will address the other comments as the relevant individual CLs come up.

But can we call the package errors?

Did we also establish consensus on the location? I think exp/errors is still the best choice. I think we can address Dave's issues if we remark in the package that the intention is to modify the respective core packages and that these packages will be removed soon after this is done.

Patch-set: 2
Reviewer: Gerrit User 5182 <5182@62eb7196-b449-3ce5-99f1-c037f21e1705>