errors: move xerrors to errors
Reasons:
1. keep diffs with core as small as possible
2. make it straightforward to use these packages
   as plug-in replacements for core in existing code.
   (particularly important for package fmt).
3. make it clear that these are plugin replacements.

As for 2, this can be done accomlished placing a copy
of these packages at the root vendor directory. It does
not seem possible to do this with modules, but it can
be achieved with modules using the following trick:

go mod vendor
cp -a $TMPGOPATH/golang.org/exp/errors vendor/
cp -a $TMPGOPATH/golang.org/exp/errors/fmt vendor/
GOPATH=&TMPGOPATH go build

This will build a project with all uses of errors.New and
fmt.Errorf replaced with their new counterparts.

Note the small change in errors. I based the
package on 1.11.1.

Change-Id: If0cf75a1514141d099168facef88ae336051f580