errors/fmt: interpret error passed to Errorf
Errors are idiomatically wrapped using Errorf by
passing them as the last argument to a format string
ending ": %s" or ": %v".

In this CL, Errorf is changed to return an error that
implements errors.Formatter, returning the last error
as the next error.

The Unwrap method is implemented in a follow-up CL.

Change-Id: Iff1ef43bf5f1cea1ad08955005e96c0a3e5d8429