Update patch set 4
Patch Set 4:

(1 comment)

I think we're going in circles on this. I'm going to +1, since I think we need someone else to resolve the deadlock one way or the other.

I do not understand how this approach can be used short of a complete, atomic replacement of the current fmt and errors packages. The vendoring-based approaches seem impractical and suited more to a GOPATH-based world, which we're leaving behind.

Patch-set: 4
Reviewer: Gerrit User 5305 <5305@62eb7196-b449-3ce5-99f1-c037f21e1705>