errors: add stack frame API
Fork of earlier work of dneil, with some
significant alterations.

Notes:
by embedding Frame in the errors, the errors
will expose the Location method. This was
a request of several people.

Frame deliberately does not implement
errors.Formatter.

Frame details includes the full function name
and file path, as this seems to be the default
for the most popular non-core errors package.
Having the full file path, at least, is useful,
as it makes it a clickable link in some shells
and IDEs.
If this is ever an issue, we could let fmt do
the rendering for Frame, and use precision
info from fmt.State to alter.

Change-Id: I0dc9f9e05c86e47a140ecfe987bdcb29c145d9be