Update patch set 1
Patch Set 1: Code-Review+2

> Patch Set 1: Code-Review-1
> 
> I'm not sure about this. I have had this feedback post lurking around for ages. I am not convinced that searching down the error chain is the right solution here. https://gist.github.com/rogpeppe/a435b57473152e3429a5a149401edacf

For now I would like to get these CLs checked into exp so people can experiment with them. Marcel is going to file an official proposal targeting Go 1.13 in a few weeks.

For what it's worth, I don't quite understand why the gist seems to praise os.IsNotExist but argue against errors.Is, when the latter seems to be just a generalization of the former that removes a significant flaw (not being able to deal with other kinds of wrapping context) and avoids having to write that kind of function over and over.

Patch-set: 1
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2