Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (3 comments)
> 
> > func (w *withStack) FormatError(p errors.Printer) error {
> > 	next := errors.Format(p, w.error)
> > 	if p.Detail() {
> > 		p.Printf("%+v", w.stack)
> > 	}
> > 	return next
> > }
> 
> Couldn't this be:
> 
> func (w *withStack) FormatError(p errors.Printer) error {
> 	if p.Detail() {
> 		p.Printf("%+v", w.stack)
> 	}
> 	return w.error
> }
> 
> This does reverse the error and the stack, but isn't that more accurate given that the *withStack may contain a stack trace captured from a different location than the wrapped error?

That would be quite awkward. The fact that withStack wraps another error is an implementation detail. withStack is typically wrapping a withMessage which is created during the same wrap function. pkg/errors could have implemented this as (at least) three different types.

As a user, I would expect that one call to a wrapping function results in one line in the output (just like pkg/errors it prints today). 

One could indeed argue that if WithStack is used in isolation the situation is different. I don't think so in the general case, but that is up to the implementers to decide. I have converted some other code and have found this kind of information decoration more often, where the wrapping error is logically the same as the one it wrapped.

Note, btw, that errors.Format would be unnecessary/ easy to code around if it were possible to find out whether or not an error will print detail without triggering detail mode. So this is partly a result of the semantics of Detail()

Patch-set: 1
Reviewer: Gerrit User 5182 <5182@62eb7196-b449-3ce5-99f1-c037f21e1705>