Update patch set 5
Patch Set 5:

> Patch Set 1:
> 
> (3 comments)
> 
> > func (w *withStack) FormatError(p errors.Printer) error {
> > 	next := errors.Format(p, w.error)
> > 	if p.Detail() {
> > 		p.Printf("%+v", w.stack)
> > 	}
> > 	return next
> > }
> 
> Couldn't this be:
> 
> func (w *withStack) FormatError(p errors.Printer) error {
> 	if p.Detail() {
> 		p.Printf("%+v", w.stack)
> 	}
> 	return w.error
> }
> 
> This does reverse the error and the stack, but isn't that more accurate given that the *withStack may contain a stack trace captured from a different location than the wrapped error?


BTW, I have removed this function. With the simplification made in this change and closing the door on other implementations, it becomes relatively easy to implement oneself:

func (w *withStack) FormatError(p errors.Printer) error {
	var next error
	if f, ok := w.error.(errors.Formatter); ok {
		next = f.FormatError(p)
	} else {
		p.Print(w.error)
	}
	if p.Detail() {
		p.Printf("%+v", w.stack)
	}
	return next
}

This does not reverse printing of the stack in the common use case where withStack is used to implement "pkg/errors".Wrap, always interleaving a withStack wit withMessage. It would indeed do so if users used "pkg/errors".WithStack directly for chains of wrapping. Arguably this is no how WithStack should be used, but this is not well defined. Theoretically, one could see if the next message has a non-empty non-detail message and only postpone printing detail if that is the case (pass a dummy printer that records prints up to a call to Detail). Removing this function at least has the benefits that we are not promoting such shenanigans.

This, btw, shows a limitation of this API: it is not possible to print the non-detail message of the next message first, then the detail of the wrapping message, and then the detail of the wrapped message. In this case that is fine, as we probably want to print the stack frame last anyway. But if there were other detail information this would not be possible. I think this is an acceptable limitation, but good to be aware off.

Patch-set: 5
Reviewer: Gerrit User 5182 <5182@62eb7196-b449-3ce5-99f1-c037f21e1705>