notary/internal/notedisk: local storage for Go notary
This is part of a design sketch for a Go module notary.
Eventually the code will live outside golang.org/x/exp.

Everything here is subject to change! Don't depend on it!

This CL implements a simple on-disk storage layer for
the data that a Go notary must maintain.

This code is not expected to be able to support a global production notary.
Instead it's meant to support experimenting with prototype
notaries. It may be useful for smaller non-global (company-internal)
notaries at some point.

Although no attempt has been made at performance,
this code can handle about 18,000 writes/second on my
MacBook Pro laptop with file system syncing turned off.
With file system syncing turned on, the performance drops
to 54 writes/second, but the benchmark is a closed-loop
single-threaded workload. A workload with multiple
parallel writers combined with appropriate batching
and pipelining should be able to cover most of the cost
of file system syncs. (Also Mac file systems sync slowly.)

Change-Id: I556a78caf86c477e97095838eee8d8abc2f2fe6c