notary/internal/note: signed note formats, keys, and operations
This is part of a design sketch for a Go module notary.
Eventually the code will live outside golang.org/x/exp.

Everything here is subject to change! Don't depend on it!

This CL implements a simple text format for the notary's
signed statements that "module M version V has file tree hash H"
as well as signed tree hashes.

Change-Id: I0d08e0943b3e2b83eace7baa059d4a05c5a330a5