notary: basic notary command
This is part of a design sketch for a Go module notary.
Eventually the code will live outside golang.org/x/exp.

Everything here is subject to change! Don't depend on it!

DO NOT REVIEW - This code is not even ready for review yet!
It's being sent as a CL only to show how notary/internal/tlog,
notary/internal/notedb, and notary/internal/note all fit together.

Change-Id: I7c1164d5c5621f744f2e00aa007e9f509cea116a