notary/internal/consistent: interface for externally consistent database
This is part of a design sketch for a Go module notary.
Eventually the code will live outside golang.org/x/exp.

Everything here is subject to change! Don't depend on it!

This code defines an interface for an externally consistent
database, such as a single-server database, or a recent
global database like Spanner or CockroachDB.

Change-Id: I7bd33711eec94b30196f998fb1003ef91e7e8d55