sumdb/internal/tkv/tkvtest: add tester and in-memory implementation
This is part of a Go checksum database implementation.
Eventually the code will live outside golang.org/x/exp.

Although the expected implementation of tkv.Storage
is a real database or key-value store, it speeds testing
considerably to avoid needing a full-blown storage system
for writing tests. This CL adds tkvtest.Mem, a trivial in-memory
implementation of tkv.Storage, for testing.

This CL also adds a preliminary tester for tkv.Storage
implementations. It is used here for testing tkvtest.Mem
but can be used for testing other implementations as well.

Change-Id: Ib70bc25a7b519fa4b685f6ce4c91b3ed04e0aa30