notary/internal/database: add in-memory DB implementation for testing
This is part of a design sketch for a Go module notary.
Eventually the code will live outside golang.org/x/exp.

Although the expected implementation of database.DB
is a real database, it speeds testing considerably to avoid
needing a full-blown database for writing tests.
This CL adds database.MemDB, a trivial in-memory
implementation of database.DB, for testing.

Change-Id: Ib70bc25a7b519fa4b685f6ce4c91b3ed04e0aa30