sumdb/internal/database: Go checksum database implementation
This is an implementation of a Go checksum database
on top of any tkv.Storage layer. It also includes a test
that can be parameterized by the tkv.Storage, to allow
end-to-end tests of the checksum database on particular
storage layers.

Eventually the code will live outside golang.org/x/exp.

Change-Id: Ifd19fd86ff16f63e534149fc0715d99537755658