notary/internal/notedb: database-backed storage for Go notary
This is part of a design sketch for a Go module notary.
Eventually the code will live outside golang.org/x/exp.
Everything here is subject to change! Don't depend on it!

This CL implements a notary storage service on top of
a consistent database such as Google Cloud Spanner.

Change-Id: Ifd19fd86ff16f63e534149fc0715d99537755658