sumdb/internal/notedb: relational database-backed storage for Go checksum database
This is part of a design sketch for a Go checksum database.
Eventually the code will live outside golang.org/x/exp.
Everything here is subject to change! Don't depend on it!

This CL implements a checksum database storage service on top of
a consistent relational database such as Google Cloud Spanner.

Change-Id: Ifd19fd86ff16f63e534149fc0715d99537755658