notary/internal/tlog: add Go signed tree head format
This is part of a design sketch for a Go module notary.
Eventually the code will live outside golang.org/x/exp.

Everything here is subject to change! Don't depend on it!

This CL adds support for formatting and parsing
Go "signed tree head" messages to be signed by
notary/internal/note.

Change-Id: Ib1f05467de439d9793d49734bcbd8bc23178b219