notary/internal/notecheck: check go.sum against notary
This is part of a design sketch for a Go module notary.
Eventually the code will live outside golang.org/x/exp.

Everything here is subject to change! Don't depend on it!

This is a demo of using the notary URL endpoints
to validate go.sum entries. The real implementation
in cmd/go would cache both the tree head and
downloaded tiles.

Change-Id: I6715cbe40e07fbab0c5cb9ab20e366195b66e6da