notary/internal/tlog: expose NodeHash
Generating internal node hashes is useful for any HashReader storage
backend that does not store all hashes, and for the update operations
that add new leafs to a tree.

Named it NodeHash instead of hashNode to mirror RecordHash.

Also fixed a minor error message bug in notecheck to fix the vet test.

Change-Id: I16bf9b0aa5ec787e1ee10e59a24ac7bbc83ea4eb