notary/internal/tlog: fix bit shifting bug when determining if an index is in a tree
The TileHashReader implementation of ReadHashes was splitting the requested
index into level and n, then bit shifting to decide whether or not the index
existing in the tree of size N. This doesn't work if the index n is 0, and
would then cause infinite looping looking for a parent tile that it knows about.

Note that this bug may exist in other parts of the code as well, but this is
the one that I noticed and was impacting me.

Change-Id: I46e27be0a0fba6d07cbabc6c9ac52cd7b24f2bc0