Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> I'm not sure why the old code was wrong but I agree the new calculation is right.

If there is an index with level=0 then it will never reach the error condition (because n<<0 will be 0 regardless of n), then just infinitely loop in the next for-loop if it can't find a parent index in the cache. The test that I added will show that infinite looping with the old code.

Patch-set: 1
Reviewer: Gerrit User 28759 <28759@62eb7196-b449-3ce5-99f1-c037f21e1705>