shiny/driver/mtldriver: add beginning of a Metal API-based driver
The OpenGL API has been deprecated on macOS as of version 10.14.¹
Building gldriver on macOS 10.14 produces warnings such as:

	# golang.org/x/exp/shiny/driver/gldriver
	cgo-gcc-prolog:117:2: warning: 'glBindVertexArray' is deprecated: first deprecated in macOS 10.14 - OpenGL API deprecated. (Define GL_SILENCE_DEPRECATION to silence these warnings) [-Wdeprecated-declarations]
	/Applications/Xcode.app/.../gl3.h:2360:21: note: 'glBindVertexArray' has been explicitly marked deprecated here
	cocoa.m:35:2: warning: 'NSOpenGLContext' is deprecated: first deprecated in macOS 10.14 - Please use Metal or MetalKit. [-Wdeprecated-declarations]
	/Applications/Xcode.app/.../NSOpenGL.h:202:12: note: 'NSOpenGLContext' has been explicitly marked deprecated here
	...

The replacement for OpenGL on macOS is the Metal API.² This change adds
the beginning of a Metal API-based shiny driver. It can be activated by
specifying the -tags=metal build tag on macOS 10.13+. For example:

	go run -tags='example metal' golang.org/x/exp/shiny/example/icongallery

The goal of this CL has been to create an MVP. As a result, the focus is
on simplicity and correctness. Performance optimizations are for later.

At this time, the Metal API is used only to present the final pixels
to the screen. All rendering is performed on the CPU via the image/draw
algorithms, and the focus so far has been on achieving correct results.
Future work is to use mtl.Buffer, mtl.Texture, etc., and do more of the
rendering work on the GPU.

From what I've observed so far, the Metal API is looking to be a great
fit for implementing the screen.Screen interface. But we'll learn more
when it's used to its full potential rather than just for blitting end
result pixels to screen.

GLFW is used for window creation, input, etc.
Package dmitri.shuralyov.com/gpu/mtl is used for Metal API access.
Helper packages for Apple's Core Animation and AppKit APIs are copied.

References:

¹ https://developer.apple.com/macos/whats-new/#deprecationofopenglandopencl
² https://developer.apple.com/metal/

DO NOT SUBMIT: This change is still a little rough.
Should do some of these things before submitting:
	- update go.mod file
	- some polish and cleanup

Change-Id: I0e02d660b776820ca499bfe7d67e47a9866d530c