Update patch set 1
Patch Set 1:

A few other suggestions that have been tossed around lately:

"sumdot" which relates to the sum.golang.org domain
"summy" which is more of a cutesy name, but kind of fits in with the theme of "proxy/summy/index"


I don't love GONOSUMDB, because we're providing a service, not just a database. But I understand that's it's an improvement from GONOVERIFY for the reasons you indicated.

Patch-set: 1
Reviewer: Gerrit User 28759 <28759@62eb7196-b449-3ce5-99f1-c037f21e1705>