Update patch set 7
Patch Set 7:

(1 comment)

> Patch Set 7:
> 
> I've addressed all the minor comments. What's left is reworking the main GLFW loop to run in a separate goroutine on the main thread and fixing the NextEvent implementation accordingly.

Minor issues look fine.

Getting the threading right is the big thing. It's not just NextEvent and the EventDeque. For example, I suspect that part of if not most of the screenImpl.NewWindow implementation should happen on "the UI thread".

IIUC, the metaphorical ball is in your court, and you'll ping the CL when I should take another look.

Patch-set: 7
Reviewer: Gerrit User 5899 <5899@62eb7196-b449-3ce5-99f1-c037f21e1705>