modgraphviz: add -colourBuild option
-colourBuild colours green the nodes chosen by MVS in the dependency graph.
This can be useful to see which nodes in the large `go mod graph` are actually
included in the build.

TODO: this still needs a lot of clean up, tests, and should work with -simple
and -pathsTo. It's currently just a hair-brained idea late at night that
turned out to work.

TODO: this CL currently copies over the relevant parts of
golang/go/src/cmd/internal/mvs/mvs.go, internal/modload/load.go, and
internal/semver/semver.go. Is there a better way to do it?

TODO: this CL does not seem to currently differentiate between major version
bumps. That is, the latest version will always be chosen irregardless of
major version bump. We should fix that in mvs.go.

Change-Id: I3d1e1c68b6cd8460e7898ede539dd6b8ccd15979