Update patch set 13
Patch Set 13:

(4 comments)

The first commit described this as "a very light piece of code that turns
go mod graph output into a .dot output that can be visualized".
That's a nice, coherent, small target. But these CLs are starting to
make it into something much bigger, a more complex, less coherent thing.
If this is something we are going to suggest that people use,
I want to make sure it stays something that is easy to understand.

We already have a tool that does queries on the "go mod graph" format,
namely golang.org/x/tools/cmd/digraph. I am not sure exactly where the
line is where it makes more sense to let digraph do analysis instead
of duplicating that analysis here, but this CL is at least close
to the line. I'm willing to proceed under the assumption that it hasn't
crossed the line yet, although I'm not 100% sure. The simpler code below
would make me feel a little more comfortable that we haven't crossed the line.

Patch-set: 13