Update patch set 13
Patch Set 13:

> We wondered whether visualization and traversal should happen in one CLI, too. A short survey didn't reveal anything that did traversal (clearly our survey was too short).
> 
> I took a peek through https://godoc.org/golang.org/x/tools/cmd/digraph. It looks like it has the traversal pieces that we need (I think --to / --focus here corresponds to allpaths there [1]).
> 
> However, a quick playthrough indicates that it's not quite emitting the kind of output we need. `allpaths` seems to just emit nodes one-per-line, rather than the adjacency matrix (`from to` on each line) that `go mod graph` emits, and `modgraphviz` expects. I think the other digraph subcommands have the same issue.
>

Digraph's current command set is for querying lists of nodes.
It does not emit graphs today. I was just saying that it probably could.
Allpaths in particular should probably emit a subgraph,
and focus would be a different operation that also emits a subgraph.
 
> This is all solveable: an --output flag that changes how things are emitted seems reasonable, for example. There's also a TODO(adonovan) in the godoc about different output formats.

I don't believe output format is the problem. The output format should be either a list of nodes (for operations that select nodes) or a graph in the digraph form (for operations that select subgraphs, of which there are zero today).

> What is your recommendation? Do you prefer me to keep this tool a simple display tool and have me work on getting digraph to play nice with modgraphviz? (with different output formats, or however else)
> 
> Or else, I'm happy to continue adding here. But you're right, this step in the chain of pipes becomes a bit fuzzy on its responsibilities.

Digraph has its own rough edges today: it's not an idiomatic Go command either as far as docs and help messages and so on; allpaths should probably emit a graph; and so on.

I think as long as you keep in mind a goal that modgraphviz be about *visualizing* the module graph and not arbitrary querying and analysis then I think it is fine to make the change in this CL (especially simplified) and make adjustments to the output to make the visualization more useful. Just keep in mind that complex querying should probably be delegated to a cleaned-up digraph.

(And of course if we do get to cleaning up digraph we'd want to loop Alan in.)

Thanks!

Patch-set: 13