Update patch set 2
Patch Set 2:

Still haven't gotten to suffix logic, but I _think_ I can just stick them in a separate list until all the normal picking/unpicking is done, then look in the original graph, find the edges it appears on, and see if any of those edges are picked (in which case we consider it also picked), and if not stick it in unpicked.

Does that seem salient?

Patch-set: 2
Reviewer: Gerrit User 26615 <26615@62eb7196-b449-3ce5-99f1-c037f21e1705>