modgraphviz: cut1
This CL causes modgraphviz to search for (solitary) edges that, when cut,
cause a part of the graph to become disconnected. For edges where the
disconnected subgraph is of significant-enough size [1], the subgraph is
displayed as a graphviz cluster. This causes graphviz to draw all the nodes in
the subgraph next to each other, and styling may be applied to the subgraph
(such as margin, border, etc).

The end result is that it's easier for users to see which nodes are "icebergs":
nodes which are depended upon in a single place, but which bring in a large
amount of under-the-water dependencies. They may be good candidates for
removal, and since they're only depended upon in a single place their removal
should be fairly easier [2].

Example with google.golang.org/api:

before: https://user-images.githubusercontent.com/3584893/62430767-0dadce00-b6de-11e9-8ec1-04f1a5b55176.png
after: https://user-images.githubusercontent.com/3584893/62430770-1900f980-b6de-11e9-957c-cc7b4fb6b765.png

(ignore the square brackets in "after" - those can be turned off)

Note that the after is a lot wider, which is a downside. On the other hand
it's now very clear to see that the following are icebergs that would be
valuable to prune (if possible):

- go.opencensus.io@v0.22.0
- google.golang.org/grpc@v1.21.1
- github.com/googleapis/gax-go/v2@v2.0.5
- google.golang.org/api@v0.7.0
- google.golang.org/appengine@v1.6.1

1: Currently set to >6 nodes in the resultant subgraph, as a simple heuristic.
2: As opposed to a dependency that is depended upon by multiple packages.

TODO:
- The output is currently non-deterministic. We should change that.
- Sometimes the current code generates subgraphs with multiple bridges. This
  _might_ be right but I suspect not. Investigate.
- Refactor to constant time with Skiena's DFS-tree approach:
    Skiena, S. (n.d.). The algorithm design manual. 2nd ed. pp.173-177.
    Skiena, S. (n.d.). The algorithm design manual. 2nd ed. pp.505-508.

Change-Id: I6090512b9e8f025edbefd735f1c00048d8d19a44