x/exp/apidiff: copy changes from x/tools/internal/apidiff
We copied golang.org/x/exp/apidiff to x/tools a few months ago in
anticipation of developing gorelease in x/tools, which would depend on
apidiff.

We've decided to develop gorelease here in x/exp instead, which means
the copy in x/tools is no longer needed. This CL copies changes made
to the copy in x/tools since it was made. Another CL will delete the
copy in x/tools.

Change-Id: Ied79138616c2f3b2f49a0ee5ca95ff3179351354