cmd/apidiff: store package path with export data
When writing export data for a package, save the path of the package.

The export data format doesn't store this information. Previously
it didn't matter, because gcexportdata.Reader.Read would return
the package that was written. But since golang.org/cl/198742, that
is no longer true.

This was apidiff's bug, not gcexportdata's: for the path argument of
Read I was passing garbage (the filename) rather than the import path.
Read's doc doesn't specify what happens if you do that.

Also, update golang.org/x/tools to latest.

Change-Id: I9d590270a9c2bb1fcc6c9071d4a2ca82c85d84d6